+++ __package__.js
@@ -1,4 +1,4 @@
 this.addScript('dom.js',['DOMImplementation','XMLSerializer']);
-this.addScript('domParser.js',['DOMHandler','DOMParser'],
+this.addScript('dom-parser.js',['DOMHandler','DOMParser'],
 		['DOMImplementation','XMLReader']);
 this.addScript('sax.js','XMLReader');
