+++ t/dom/require.t.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
 require('proof')(1, function (ok) {
-  ok(require('../../domParser'), 'require');
+  ok(require('../../dom-parser'), 'require');
 });
