+++ test/parse/test-define.js
@@ -1,3 +1,3 @@
 console.log(__dirname)
 exports.XMLSerializer = require('../../dom').XMLSerializer ;
-exports.DOMParser = require('../../domParser').DOMParser;
+exports.DOMParser = require('../../dom-parser').DOMParser;
